+++ script/src/lib.rs
@@ -1,8 +1,6 @@
 //! CKB component to run the type/lock scripts.
 pub mod cost_model;
 mod error;
-mod initial_load;
-mod program_cache;
 mod scheduler;
 mod syscalls;
 mod type_id;
@@ -11,7 +9,6 @@ mod verify;
 mod verify_env;
 
 pub use crate::error::{ScriptError, TransactionScriptError};
-pub use crate::initial_load::{InitialProgramLoadLimit, InitialProgramLoadReceipt};
 pub use crate::scheduler::{ROOT_VM_ID, Scheduler};
 pub use crate::syscalls::generator::generate_ckb_syscalls;
 pub use crate::types::{